:recycle: :card_file_box: Modify score field to decimal Comment out unnecessary logs

+++ src/database/migrations/20230701094008-create-portfolio.js
@@ -16,7 +16,7 @@ export async function up(queryInterface, Sequelize) {
       allowNull: false,
     },
     score: {
-      type: Sequelize.INTEGER,
+      type: Sequelize.DECIMAL(25, 10),
       defaultValue: 0,
     },
     subCategoryId: {
